+++ CHANGES.rst
@@ -9,7 +9,9 @@ Changelog
   Closes #12.
 * Differentiate packages names and IDs. Closes #11.
 * Sort list of outdated packages by lower-cased package names first.
-* Add update CLI field for each outdated packages in JSON output.
+* Add ``upgrade_cli`` field for each outdated packages in JSON output.
+* Allow user to choose rendering of ``upgrade_cli`` field to either one-liner,
+  fragments or BitBar format. Closes #14.
 * Include errors reported by each manager in JSON output of ``mpm outdated``
   command.
 * Fix parsing of multiple versions of ``cask`` installed packages.